requirements.txt
colorama==0.3.7
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
djoser==0.5.4
drf-pdf==0.2.0
freeze==1.0.10
gunicorn==19.7.1
httpie==0.9.9
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
Pygments==2.2.0
PyJWT==1.4.2
pytz==2017.2
requests==2.13.0
six==1.10.0
